pyproject.toml
[project]
name = "phase1-track3-lab1-advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "pandas>=2.2",
    "pydantic>=2.7",
    "pytest>=9.0.3",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "typer>=0.12",
]
